separate out lock events to avoid circular import
